# perf(astro/assets): avoid downloading original image when using cache