--- a/.changeset/purple-ears-sneeze.md
+++ b/.changeset/purple-ears-sneeze.md
@@ -1,0 +1,5 @@
+---
+'astro': patch
+---
+
+perf(assets): avoid downloading original image when using cache